requirements.txt
absl-py==0.12.0
asgiref==3.3.4
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
Django==3.0.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
djongo==1.3.4
flatbuffers==1.12
gast==0.3.3
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
Pillow==8.2.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.3
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sqlparse==0.2.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
